\name{my.print.ctab}
\alias{my.print.ctab}
\title{my.print.ctab}
\usage{
  my.print.ctab(x, dec.places = x$dec.places,
    addmargins = x$addmargins, all.NA = TRUE, ...)
}
\arguments{
  \item{x}{An object output from ctab()}

  \item{dec.places}{Decimal places to print in output}

  \item{addmargins}{Should margins be calculated?}

  \item{all.NA}{Boolean, should NA values be recognized or
  skipped}

  \item{...}{}
}
\description{
  my.print.ctab
}

