\name{vector.diff}
\alias{vector.diff}
\title{vector.diff}
\usage{
  vector.diff(x, y)
}
\arguments{
  \item{x}{a numeric or integer vector}

  \item{y}{a numeric or integer vector of length x}
}
\value{
  A vector of length x with the absolute pairwise
  differences between x and y
}
\description{
  Returns the vector of absolute differences of two vectors
}
\author{
  Mark Huberty
}

